Having a map working locally. Showing categories and number of bikes.  Helper function are in app_utils.
